[REF] Refactoring user information
Odoo crypt password now so every time we try to get it, we have an empty string. With old code every user password would have been overwritten. With this refactoring we use an admin/manager account to write data for other users but force on creation the right user reference.